Own modules for fragment and level-key classes
